+++ tests/test_strings/test_format_time.py
@@ -4,19 +4,21 @@ from rics.strings import format_perf_counter, format_seconds
 
 
 class Base:
-    @staticmethod
-    def test_positive(t, expected):
-        assert format_seconds(t) == expected
-        assert format_perf_counter(0, end=t) == expected
+    full: bool
 
-    @staticmethod
-    def test_negative(t, expected):
+    @classmethod
+    def test_positive(cls, t, expected):
+        assert format_seconds(t, full=cls.full) == expected
+        assert format_perf_counter(0, end=t, full=cls.full) == expected
+
+    @classmethod
+    def test_negative(cls, t, expected):
         t = -t
         expected = f"-{expected}"
 
         with pytest.raises(ValueError):
             format_seconds(t)
-        assert format_seconds(t, allow_negative=True) == expected
+        assert format_seconds(t, allow_negative=True, full=cls.full) == expected
 
 
 @pytest.mark.parametrize(
@@ -36,7 +38,7 @@ class Base:
     ],
 )
 class TestSubSecond(Base):
-    pass
+    full = True  # Doesn't matter
 
 
 @pytest.mark.parametrize(
@@ -51,7 +53,7 @@ class TestSubSecond(Base):
     ],
 )
 class TestSecond(Base):
-    pass
+    full = True  # Doesn't matter
 
 
 @pytest.mark.parametrize(
@@ -63,21 +65,50 @@ class TestSecond(Base):
         (7794363.9, "90d 5h 6m 4s"),
     ],
 )
-class TestClock(Base):
-    pass
+class TestClockFull(Base):
+    full = True
 
 
 @pytest.mark.parametrize(
     "t, expected",
     [
-        (59.99, "60.0s"),
-        (60.00, "60.0s"),
-        (60.01, "1m"),
-        (61.0, "1m 1s"),
+        (123.3, "2m 3s"),
+        (18003.9, "5h"),
+        (18363.9, "5h 6m"),
+        (7794363.9, "90d 5h 6m"),
     ],
 )
+class TestClock(Base):
+    full = False
+
+
+TRANSITION_CASES = [(59.99, "60.0s"), (60.00, "60.0s"), (60.01, "1m"), (61.0, "1m 1s")]
+
+
+@pytest.mark.parametrize("t, expected", TRANSITION_CASES)
+class TestClockTransitionFull(Base):
+    full = True
+
+
+@pytest.mark.parametrize("t, expected", TRANSITION_CASES)
 class TestClockTransition(Base):
-    pass
+    full = False
+
+
+@pytest.mark.parametrize(
+    "t, expected",
+    [
+        (119.49, "1m 59s"),
+        (119.50, "2m"),
+        (119.51, "2m"),
+        (3599.50, "1h"),
+        (3599.49, "59m 59s"),
+        (3659.50, "1h 1m"),
+        (3659.49, "1h 0m 59s"),
+    ],
+)
+class TestBoundaryClockFull(Base):
+    full = True
 
 
 @pytest.mark.parametrize(
@@ -93,4 +124,4 @@ class TestClockTransition(Base):
     ],
 )
 class TestBoundaryClock(Base):
-    pass
+    full = False